+++ balances/index.js
@@ -1,11 +1,13 @@
 const balanceFromTokens = require('./balance_from_tokens');
 const getAccountingReport = require('./get_accounting_report');
 const getBalance = require('./get_balance');
+const getDetailedBalance = require('./get_detailed_balance');
 const getLiquidity = require('./get_liquidity');
 
 module.exports = {
   balanceFromTokens,
   getAccountingReport,
   getBalance,
+  getDetailedBalance,
   getLiquidity,
 };